fix(lens): one formatUptime() for the one daemon clock
Hoist a single formatUptime(ms) into useDaemonStatus.ts; the rail and the
kaval dialog both import it and delete their local uptime formatters.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-4, raised by lowy). Not pushed or merged.
